fix(ci): unbreak macOS build and Windows vet
Two platform-specific CI failures surfaced once the Actions billing
issue cleared and the full matrix ran again:

- macOS `go build` failed: go.mod pinned `go 1.25.11`, a patch newer
  than the runner's 1.25.10, and setup-go's GOTOOLCHAIN=local blocked
  auto-download. Relax to `go 1.25.0` so any 1.25.x toolchain works.

- Windows `go vet` failed: archive_test.go called unix-only
  syscall.Mkfifo unconditionally, so the test package wouldn't compile
  on Windows (a runtime GOOS skip can't help). Move the FIFO test to
  archive_unix_test.go behind //go:build unix.

Verified: go vet ./... (native + GOOS=windows), GOOS=darwin go build,
and full go test ./... all pass.
