+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/polius/fsend
 
-go 1.25.11
+go 1.25.0
 
 require (
 	github.com/adrg/xdg v0.5.3
